refactor(storage): remove standalone blob backend seam
