+++ packages/@openmaic/storage/README.md
@@ -51,9 +51,12 @@ a browser.
   defend against does not arise. Images, audio and video share one id space —
   the medium is a `mime` column, not a partition.
   `put` **always allocates a new id**, so its successful return values and
-  branches do not reveal whether the bytes were already present; the
-  content-addressed blob backend beneath, whose refs would disclose exactly
-  that, is internal and unexported. Resource-accounting channels remain:
+  branches do not reveal whether the bytes were already present. The browser
+  registry embeds its `blobs` table in the same database because reference
+  counting, byte writes, and reclamation must share one transaction. This is
+  not a replaceable browser-side blob backend; a replaceable blob interface is
+  a server-backend concern (delivery plan part 4), where consistency is enforced
+  server-side. Resource-accounting channels remain:
   quota errors, storage estimates, and server billing or metering can disclose
   existence, so server deployments must budget them per principal. Object URLs
   are minted per id, not shared per `contentHash`: sharing would let a holder of
@@ -109,9 +112,9 @@ a browser.
 
 ## Upgrading from 0.1.x
 
-Version 0.2.0 removes `BrowserAssetProvider` from the package entry point. The
-outward asset API is now `BrowserAssetStore`, whose refs are allocated ids and
-whose data lives in the new `maic-asset-pool` database.
+Version 0.2.0 removes `BrowserAssetProvider` outright; it no longer ships. The
+asset API is now `BrowserAssetStore`, whose refs are allocated ids and whose
+data lives in the new `maic-asset-pool` database.
 
 Reusing a custom `dbName` created by a 0.1.x provider raises an explicit
 legacy-schema error rather than corrupting data or operating only partially.
@@ -132,10 +135,8 @@ Each primitive has one implementation-agnostic contract suite
 `test/runtime-contract.ts`).
 Every backend is proven by running the same suite against it, so browser, HTTP,
 and PostgreSQL implementations cannot silently diverge from a primitive's
-semantics. The asset layer has two, because its two layers have deliberately
-opposite semantics: `test/asset-contract.ts` for the outward allocated-id store
-(identical bytes never share an id) and `test/asset-blob-contract.ts` for the
-internal blob layer (identical bytes always share a key). Outward-store
+semantics. Assets use the single `test/asset-contract.ts` suite for the
+allocated-id store: identical bytes never share a caller-visible id. Asset
 backends must let the suite temporarily instrument the production allocation
 source while constructing the store through the same factory used by every
 other contract test. This proves that every successful `put` consumes exactly
@@ -145,10 +146,8 @@ adding a caller-configurable allocation path.
 ## Roadmap
 
 - [x] `KVStore` + browser backend; zustand `persist` adapter
-- [x] `StorageProvider` (in `@openmaic/dsl`) + browser content-addressed blob
-      backend
-- [x] asset registry: allocated `AssetId` over the blob layer, browser
-      `BrowserAssetStore` (#1007)
+- [x] `StorageProvider` (in `@openmaic/dsl`) + browser asset registry: allocated
+      `AssetId` over an embedded byte table in `BrowserAssetStore` (#1007)
 - [x] implementation-agnostic contract suites
 - [x] `DocumentStore` (aggregate ↔ normalized adapter, migrate-on-read via the
       DSL migration registry, validation gate) + browser backend
@@ -176,9 +175,10 @@ adding a caller-configurable allocation path.
 - [x] `KVStore` (`account`) HTTP backend + HTTP contract
 - [ ] `KVStore` server-side reference backend and reference-server route
 - [ ] asset server backend — registry (principal column, server-derived) +
-      blob store, over the global resource pool model (#1007). It must validate
-      or allowlist content types before serving bytes rather than reflecting
-      cross-principal metadata into response content types
+      replaceable blob storage, over the global resource pool model (#1007),
+      with consistency enforced server-side. It must validate or allowlist
+      content types before serving bytes rather than reflecting cross-principal
+      metadata into response content types
 - [ ] asset manifest: the one enumeration of "which `AssetId`s does this course
       reference?" the export paths converge on (#1007)
