--- packages/@openmaic/storage/src/asset/browser.ts
@@ -1,196 +1,0 @@
-/**
- * **Internal.** The standalone content-addressed blob backend — bytes in
- * IndexedDB, keyed by `sha256-<hex>` over their own content, resolved as a
- * `blob:` object URL.
- *
- * This is not the package's asset seam and is deliberately not exported:
- * consumers get {@link BrowserAssetStore}, which layers an allocated-id
- * registry over this addressing scheme. The distinction matters, because a
- * content-addressed ref *is* a statement about the bytes — putting the same
- * bytes twice yields the same ref, which is a de-duplication signal an outward
- * API must not emit. That property is fine one layer down, where a hash is a
- * private storage key nothing outside the package holds; it is not fine as a
- * reference a document embeds.
- *
- * Kept as its own type because "map bytes to a stable storage key" is the
- * pluggable half of the design — the layer an S3-compatible endpoint or a CDN
- * would replace behind the same interface, with the registry above it
- * unchanged.
- */
-import type { AssetMeta, BinaryBlob } from '@openmaic/dsl';
-import { contentHashOf, ObjectUrlCache, type BlobStore, type ContentHash } from './blob.js';
-
-export interface BrowserAssetProviderOptions {
-  /** IndexedDB factory. Defaults to the ambient `indexedDB`. Injectable for tests. */
-  indexedDB?: IDBFactory;
-  /** Database name. Defaults to `maic-assets`. */
-  dbName?: string;
-}
-
-const STORE = 'assets';
-
-function incompatibleSchemaError(dbName: string): Error {
-  return new Error(
-    `IndexedDB database "${dbName}" has an incompatible asset schema. ` +
-      'It may have been created by BrowserAssetStore as a registry pool or by another incompatible version. ' +
-      'Use a fresh dbName, or wait for or request the explicit one-time import helper.',
-  );
-}
-
-interface StoredAsset {
-  bytes: ArrayBuffer;
-  contentType: string;
-}
-
-export class BrowserAssetProvider implements BlobStore {
-  private readonly idb: IDBFactory;
-  private readonly dbName: string;
-  private dbPromise?: Promise<IDBDatabase>;
-  private readonly urls = new ObjectUrlCache<ContentHash>((left, right) => left === right);
-  private closed = false;
-  private closePromise?: Promise<void>;
-
-  constructor(options: BrowserAssetProviderOptions = {}) {
-    this.idb = options.indexedDB ?? globalThis.indexedDB;
-    this.dbName = options.dbName ?? 'maic-assets';
-  }
-
-  private assertOpen(): void {
-    if (this.closed) throw new Error('BrowserAssetProvider is closed.');
-  }
-
-  private openDb(): Promise<IDBDatabase> {
-    this.assertOpen();
-    // Do NOT cache a rejected open: a transient failure (private-mode IDB, a
-    // one-off VersionError) would otherwise brick the provider for the whole
-    // session. Clear the memo on failure so the next call retries.
-    this.dbPromise ??= new Promise<IDBDatabase>((resolve, reject) => {
-      const req = this.idb.open(this.dbName, 1);
-      req.onupgradeneeded = () => {
-        if (!req.result.objectStoreNames.contains(STORE)) req.result.createObjectStore(STORE);
-      };
-      req.onsuccess = () => {
-        const db = req.result;
-        try {
-          const hasOnlyExpectedStore =
-            db.objectStoreNames.length === 1 && db.objectStoreNames.contains(STORE);
-          const hasNoIndexes =
-            hasOnlyExpectedStore &&
-            db.transaction(STORE, 'readonly').objectStore(STORE).indexNames.length === 0;
-          if (!hasOnlyExpectedStore || !hasNoIndexes) {
-            db.close();
-            reject(incompatibleSchemaError(this.dbName));
-            return;
-          }
-          resolve(db);
-        } catch {
-          db.close();
-          reject(incompatibleSchemaError(this.dbName));
-        }
-      };
-      req.onerror = () => reject(req.error);
-    }).catch((err) => {
-      this.dbPromise = undefined;
-      throw err;
-    });
-    return this.dbPromise;
-  }
-
-  private async tx<T>(
-    mode: IDBTransactionMode,
-    run: (store: IDBObjectStore) => IDBRequest<T>,
-  ): Promise<T> {
-    const db = await this.openDb();
-    return new Promise<T>((resolve, reject) => {
-      const transaction = db.transaction(STORE, mode);
-      const req = run(transaction.objectStore(STORE));
-      let result: T;
-      req.onsuccess = () => {
-        result = req.result;
-      };
-      // Resolve on commit, not on the request success: a write that succeeds
-      // as a request can still abort at commit (e.g. QuotaExceededError), and
-      // reporting that as success would claim durability the store never gave.
-      transaction.oncomplete = () => resolve(result);
-      transaction.onerror = () => reject(transaction.error ?? req.error);
-      transaction.onabort = () => reject(transaction.error ?? req.error);
-    });
-  }
-
-  put = async (data: BinaryBlob, meta?: AssetMeta): Promise<ContentHash> => {
-    this.assertOpen();
-    const { contentHash, bytes } = await contentHashOf(data);
-    const asset: StoredAsset = { bytes, contentType: meta?.contentType ?? data.type ?? '' };
-    await this.tx('readwrite', (store) => store.put(asset, contentHash));
-    // A re-put with the same bytes but different metadata (e.g. a corrected
-    // contentType) overwrites the stored asset; retire the current object URL
-    // for this key so future resolve() calls reflect the latest write. The old
-    // URL remains a live snapshot until release for this key or provider close.
-    await this.urls.invalidate(contentHash);
-    return contentHash;
-  };
-
-  /**
-   * Resolve a content hash to a `blob:` URL, or `null` when it is absent.
-   *
-   * A resolve that passes the open check but races `close` may return a
-   * non-null URL that `close` has already revoked; callers tearing down should
-   * not race resolve against close.
-   */
-  resolve = async (ref: ContentHash): Promise<string | null> => {
-    this.assertOpen();
-    return this.urls.resolve(ref, ref, () => this.readAsUrl(ref));
-  };
-
-  private async readAsUrl(
-    ref: ContentHash,
-  ): Promise<{ identity: ContentHash; url: string } | null> {
-    const asset = await this.tx<StoredAsset | undefined>('readonly', (store) => store.get(ref));
-    if (!asset) return null;
-    // Mint only after the readonly transaction commits, so an aborted read
-    // cannot leak an object URL that no caller ever receives.
-    return {
-      identity: ref,
-      url: URL.createObjectURL(new Blob([asset.bytes], { type: asset.contentType })),
-    };
-  }
-
-  remove = async (ref: ContentHash): Promise<void> => {
-    this.assertOpen();
-    await this.tx('readwrite', (store) => store.delete(ref));
-    await this.urls.invalidate(ref);
-  };
-
-  /**
-   * Revoke this provider's current and retired snapshots for `ref`.
-   *
-   * This is an owner-level operation: it is safe only when the caller owns
-   * every consumer of every URL returned for that ref by this provider.
-   */
-  release = async (ref: ContentHash): Promise<void> => {
-    this.assertOpen();
-    await this.openDb();
-    await this.urls.release(ref);
-  };
-
-  /**
-   * Close this provider and reclaim all of its snapshots.
-   *
-   * This is an owner-level operation: it is safe only when the caller owns
-   * every consumer of every URL returned by this provider. If an unawaited
-   * `release` is already in flight, its last URL may be revoked one microtask
-   * after this call resolves.
-   */
-  close = (): Promise<void> => {
-    if (this.closePromise) return this.closePromise;
-    this.closed = true;
-    const closeDb = this.dbPromise
-      ? this.dbPromise.then(
-          (db) => db.close(),
-          () => undefined,
-        )
-      : Promise.resolve();
-    this.closePromise = Promise.all([this.urls.close(), closeDb]).then(() => undefined);
-    return this.closePromise;
-  };
-}
