--- packages/@openmaic/storage/test/asset-blob-contract.ts
@@ -1,119 +1,0 @@
-// Contract for the INTERNAL blob layer: `contentHash -> bytes`, where the key
-// is derived from the bytes themselves. This is the layer an object-storage /
-// CDN backend would replace, so its semantics get their own suite — and they
-// are deliberately *not* the semantics of the outward asset store, which
-// allocates ids (see `asset-contract.ts`). Content addressing is safe here
-// precisely because a hash never leaves the package.
-import { describe, expect, test, vi } from 'vitest';
-import type { BlobStore, ContentHash } from '../src/asset/blob.js';
-
-type ReadUrl = (url: string) => Promise<Uint8Array>;
-
-const bytes = (s: string): Uint8Array => new TextEncoder().encode(s);
-// Build the Blob from the string directly (a string is a valid BlobPart). A
-// `Uint8Array` BlobPart trips TS 5.7+'s `Uint8Array<ArrayBufferLike>` vs
-// `ArrayBufferView<ArrayBuffer>` narrowing under the root tsconfig; the bytes
-// are UTF-8 either way, so `bytes(s)` stays the source of truth for comparison.
-const blob = (s: string, type = 'text/plain'): Blob => new Blob([s], { type });
-
-export function runBlobStoreContract(
-  name: string,
-  makeProvider: () => BlobStore,
-  readUrl: ReadUrl,
-): void {
-  describe(`blob layer contract: ${name}`, () => {
-    test('put returns a non-empty key', async () => {
-      const p = makeProvider();
-      const key = await p.put(blob('hello'));
-      expect(typeof key).toBe('string');
-      expect(key.length).toBeGreaterThan(0);
-    });
-
-    test('is content-addressed: identical bytes yield the same key', async () => {
-      const p = makeProvider();
-      const a = await p.put(blob('same-content'));
-      const b = await p.put(blob('same-content'));
-      expect(a).toBe(b);
-    });
-
-    test('distinct bytes yield distinct keys', async () => {
-      const p = makeProvider();
-      const a = await p.put(blob('one'));
-      const b = await p.put(blob('two'));
-      expect(a).not.toBe(b);
-    });
-
-    test('resolve yields a URL whose bytes equal the stored blob', async () => {
-      const p = makeProvider();
-      const key = await p.put(blob('round-trip me'));
-      const url = await p.resolve(key);
-      expect(url).not.toBeNull();
-      expect(await readUrl(url!)).toEqual(bytes('round-trip me'));
-    });
-
-    test('concurrent resolve of the same key yields one shared URL', async () => {
-      const p = makeProvider();
-      const key = await p.put(blob('shared'));
-      // Two in-flight resolves must not each mint a URL (the second would orphan
-      // the first, which only `remove` could ever revoke). They share one.
-      const [a, b] = await Promise.all([p.resolve(key), p.resolve(key)]);
-      expect(a).not.toBeNull();
-      expect(a).toBe(b);
-    });
-
-    test('resolve returns null for an unknown key', async () => {
-      const p = makeProvider();
-      expect(await p.resolve('sha256-deadbeef' as ContentHash)).toBeNull();
-    });
-
-    test('resolve returns null after remove', async () => {
-      const p = makeProvider();
-      const key = await p.put(blob('temporary'));
-      await p.remove(key);
-      expect(await p.resolve(key)).toBeNull();
-    });
-
-    test('release reclaims cached snapshots without removing stored bytes', async () => {
-      const p = makeProvider();
-      const key = await p.put(blob('release me'));
-      const first = await p.resolve(key);
-      expect(first).not.toBeNull();
-
-      await p.release(key);
-
-      const second = await p.resolve(key);
-      expect(second).not.toBeNull();
-      expect(second).not.toBe(first);
-      expect(await readUrl(second!)).toEqual(bytes('release me'));
-      await p.close();
-    });
-
-    test('close is idempotent and every later operation fails loudly', async () => {
-      const p = makeProvider();
-      const key = await p.put(blob('close me'));
-      await p.resolve(key);
-
-      await expect(p.close()).resolves.toBeUndefined();
-      await expect(p.close()).resolves.toBeUndefined();
-      await expect(p.put(blob('later'))).rejects.toThrow(/closed/i);
-      await expect(p.resolve(key)).rejects.toThrow(/closed/i);
-      await expect(p.remove(key)).rejects.toThrow(/closed/i);
-      await expect(p.release(key)).rejects.toThrow(/closed/i);
-    });
-
-    test('close closes an opened IndexedDB connection', async () => {
-      const p = makeProvider();
-      await p.put(blob('open the database'));
-      const connection = await (p as unknown as { dbPromise?: Promise<IDBDatabase> }).dbPromise;
-      if (!connection) throw new Error('expected an opened IndexedDB connection');
-      const prototype = Object.getPrototypeOf(connection) as IDBDatabase;
-      const close = vi.spyOn(prototype, 'close');
-      try {
-        await p.close();
-        expect(close).toHaveBeenCalledTimes(1);
-      } finally {
-        close.mockRestore();
-      }
-    });
-  });
-}
